third_party/blink/web_tests/fast/filesystem/async-operations-expected.txt

Test making multiple async FileSystem operations.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

requested FileSystem.
Got FileSystem:file__0:Temporary
Starting async test stage 1.
Starting async test stage 2.
Verifying the FileSystem status.
PASS "/a,/b,/c,/d2,/e,/f" is "/a,/b,/c,/d2,/e,/f"
PASS dirsCount is 3
PASS successfullyParsed is true

TEST COMPLETE
